flask_postgres_k8s/app/requirements.txt
Flask
psycopg2-binary
prometheus-flask-exporter==0.22.4